Created Sample File From Good Racks LLC
This file was created completely by me for my startup Good Racks LLC. It contains the class used by our automated production system specifically for interfacing with one of our machines that help produce our product. The other classes and files for this project are not included because that is all proprietary for the company, however we removed the more proprietary portions of this code and are comfortable sharing what is left.